new api kwjr_get()

+++ c_prot/src/c_prot_modbus_master.c
@@ -3119,7 +3119,7 @@ PRIVATE cell_control_t *get_cell_control(
         json_object_set_new(jn_type, saddress, json_integer(0));
     }
     json_t *jn_value = json_object_get(jn_type, saddress);
-    cell_control_t *cell_control = (cell_control_t *)json_integer_value_pointer(jn_value);
+    cell_control_t *cell_control = (cell_control_t *)(size_t)json_integer_value_pointer(jn_value);
     return cell_control;
 }
